% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniConSig_usingPreCal_functionalized.R
\name{cal_uniConSig}
\alias{cal_uniConSig}
\title{calculate uniConSig using pre-calculated files}
\usage{
cal_uniConSig(trList, preCal = preCal.data)
}
\arguments{
\item{trList}{A vector containing all the training gene ids.}

\item{preCal}{Default using a small subset of the whole pre-calculation file. To do full calculation, see the examples}
}
\value{
A matrix with 3 columns: gene ID, gene Name, and uniConSig score
}
\description{
calculate the functional associations between a single gene and a training gene set stored in trList.
}
\examples{
#For customized training list:
trList.call<-1:10 #read customized training list
result<-cal_uniConSig(trList.call) #calculate uniConSig
#To do full calculation:
#preCal.local<-get_data_uniConSigPreCal()
#load(preCal.local)
#result<-cal_uniConSig(trList.call,preCal=preCal.data.all)
#head(result) #examine the result
}
