% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniConSig_usingPreCal_functionalized.R
\name{get_data_uniConSigPreCal}
\alias{get_data_uniConSigPreCal}
\title{Fetch pre-calculated data from github for uniConSig calculation}
\usage{
get_data_uniConSigPreCal(file.data = "preCal.data.all.RData",
  dir.local = "./data.uniConSigPreCal/")
}
\arguments{
\item{file.data}{A string which specifies the name of the data file to be fetched from github. Default is preCal.data.all.RData}

\item{dir.local}{A string which specifies the name of the directory where the data file will be saved. Need "/" at the end of the string.}
}
\value{
The name of the downloaded file with the path to it.
}
\description{
Download the pre-calculated data from github. Returns the name of the downloaded file with the path to it.
}
\examples{
#For the data required in uniConSig calculation:
#preCal.local<-get_data_uniConSigPreCal()
#load(preCal.local)
}
