% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniConSig_usingPreCal_functionalized.R
\docType{data}
\name{pathway.c2cp}
\alias{pathway.c2cp}
\title{Pathway gene sets of C2CP from MSigDB}
\format{A list of list, elements are entrez gene ids. Names of each sublist is the pathway's name}
\usage{
data(pathway.c2cp)
}
\description{
Compiled gene sets from MSigdDB c2cp
}
\author{
Xu Chi (original data from MSigDB c2cp) \email{xuc13@pitt.edu}
}
\keyword{datasets}
