% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniConSig_usingPreCal_functionalized.R
\docType{data}
\name{pathway.hallmark}
\alias{pathway.hallmark}
\title{Pathway gene sets of HALLMARK from MSigDB}
\format{A list of list, elements are entrez gene ids. Names of each sublist is the pathway's name}
\usage{
data(pathway.hallmark)
}
\description{
Compiled gene sets from MSigDB hallmark
}
\author{
Xu Chi (original data from MSigDB c2cp) \email{xuc13@pitt.edu}
}
\keyword{datasets}
