% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniConSig_usingPreCal_functionalized.R
\name{perm_weightedKS_ofCSEA}
\alias{perm_weightedKS_ofCSEA}
\title{Calculate the permutations of random on-set genes}
\usage{
perm_weightedKS_ofCSEA(weights, numOnList, nPermu = 1000)
}
\arguments{
\item{weights}{A vector of weights. The name of each value in the vector is it's Entrez gene ID}

\item{numOnList}{An positive integer, indicating the number of genes to be randomly selected}

\item{nPermu}{Number of permutations, default 1000}
}
\value{
A numeric vector of enrichment scores of randomly selected genes
}
\description{
For a certain weighted gene list(weights), randomly select a certain number(numOnList) of genes to be on-set genes, and calculate ES
}
\examples{
weight.call<-c(0.1,0.2,0.5,0.7,0.9,1)
names(weight.call)<-1:6
test.tmp<-perm_weightedKS_ofCSEA(weight.call,6)
}
